requirements.txt
adbutils==0.8.2
yagmail==0.14.260
aiocontextvars==0.2.2
allure-pytest==2.8.18
allure-python-commons==2.8.18
numpy==1.19.5
scikit-build==0.11.1
opencv-python==4.5.1.48
py==1.9.0
pytest==6.1.0
scikit-image==0.17.2
scipy==1.5.4
uiautomator2==2.12.0
urllib3==1.25.11
zipp==3.2.0
pyserial==3.5
delayed_assert==0.3.5
loguru==0.5.3

